backend/requirements.txt
fastapi
uvicorn
web3
py-solc-x
python-dotenv
mysql-connector-python